pyproject.toml
[project]
name = "job-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=1.12.0",
    "groq>=0.30.0",
    "mcp[cli]>=1.12.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "transformers>=4.53.2",
]
